requirements.txt
requests
beautifulsoup4
mcp
pydantic
python-dotenv
